Разбиваю проект на модули
