+++ js/data.js
@@ -1,8 +1,8 @@
 import {getRandomNumberInclusive, getRandomNumberFloat} from './utils.js';
-const initialCoordinates =  {
-  lat: 35.68950,
-  lng: 139.69171,
-};
+// const initialCoordinates =  {
+//   lat: 35.68950,
+//   lng: 139.69171,
+// };
 const mockOffer = {
   title: 'Три поросенка',
   address: '',
@@ -79,4 +79,4 @@ const getOfferTypes = () => ({
   hotel: 'Отель',
 });
 
-export {getAuthor, getOffer, getLocation, getOfferTypes, initialCoordinates};
+export {getAuthor, getOffer, getLocation, getOfferTypes};
